requirements.txt
flask
python-dotenv
langchain-community
langchain-text-splitters
langgraph
langchain
langchain-groq
pinecone-client
langchain-pinecone
sentence-transformers
langchain-huggingface
